+++ src/main/java/shadows/placebo/json/PlaceboJsonReloadListener.java
@@ -1,5 +1,6 @@
 package shadows.placebo.json;
 
+import java.lang.ref.WeakReference;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
@@ -25,12 +26,25 @@ import net.minecraft.server.packs.resources.SimpleJsonResourceReloadListener;
 import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.crafting.CraftingHelper;
+import net.minecraftforge.common.crafting.conditions.ICondition;
+import net.minecraftforge.common.crafting.conditions.ICondition.IContext;
+import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.event.OnDatapackSyncEvent;
 import shadows.placebo.Placebo;
 import shadows.placebo.json.PlaceboJsonReloadListener.TypeKeyed;
 import shadows.placebo.network.PacketDistro;
 import shadows.placebo.packets.ReloadListenerPacket;
 
+/**
+ * A Placebo JSON Reload Listener is a big pile of boilerplate for registering reload listeners efficiently.<br>
+ * To utilize this class, subclass it, and provide the appropriate constructor parameters.<br>
+ * You will provide your serializers via {@link PlaceboJsonReloadListener#registerBuiltinSerializers()}.<br>
+ * You will then need to register it via {@link PlaceboJsonReloadListener#registerToBus()}.<br>
+ * 
+ * From then on, loading of files, condition checks, network sync, and everything else is automatically handled.
+ *
+ * @param <V> The base type of objects stored in this reload listener.
+ */
 public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends SimpleJsonResourceReloadListener {
 
 	private static final Map<String, PlaceboJsonReloadListener<?>> SYNC_REGISTRY = new HashMap<>();
@@ -42,11 +56,19 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 	protected final boolean synced;
 	protected final boolean subtypes;
 	protected final BiMap<ResourceLocation, SerializerBuilder<V>.Serializer> serializers = HashBiMap.create();
+	private WeakReference<ICondition.IContext> context;
 
 	protected Map<ResourceLocation, V> registry = ImmutableMap.of();
 
 	private final Map<ResourceLocation, V> staged = new HashMap<>();
 
+	/**
+	 * Constructs a new JSON reload listener.  All parameters will be saved finally in the object.
+	 * @param logger The logger used by this listener for all relevant messages.
+	 * @param path The datapack path used by this listener for loading files.
+	 * @param synced If this listener will be synced over the network.  You must also call {@link PlaceboJsonReloadListener#registerForSync(PlaceboJsonReloadListener)}
+	 * @param subtypes If this listener supports subtyped objects (and the "type" key on top-level objects).
+	 */
 	public PlaceboJsonReloadListener(Logger logger, String path, boolean synced, boolean subtypes) {
 		super(new GsonBuilder().setLenient().create(), path);
 		this.logger = logger;
@@ -55,11 +77,6 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		this.subtypes = subtypes;
 		this.registerBuiltinSerializers();
 		if (this.serializers.isEmpty()) throw new RuntimeException("Attempted to create a json reload listener for " + path + " with no top-level serializers!");
-		if (synced) {
-			if (SYNC_REGISTRY.containsKey(path)) throw new RuntimeException("Attempted to create a synced json reload listener for " + path + " but one already exists!");
-			SYNC_REGISTRY.put(path, this);
-			MinecraftForge.EVENT_BUS.addListener(this::sync);
-		}
 	}
 
 	@Override
@@ -67,7 +84,7 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		this.beginReload();
 		objects.forEach((key, ele) -> {
 			try {
-				if (checkAndLogEmpty(ele, key, this.path, this.logger) && checkConditions(ele, key, this.path, this.logger)) {
+				if (checkAndLogEmpty(ele, key, this.path, this.logger) && checkConditions(ele, key, this.path, this.logger, this.getContext())) {
 					JsonObject obj = ele.getAsJsonObject();
 					SerializerBuilder<V>.Serializer serializer;
 					if (this.subtypes && obj.has("type")) {
@@ -115,6 +132,12 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		this.logger.info("Registered {} {}.", this.registry.size(), this.path);
 	}
 
+	/**
+	 * Register a serializer to this listener.  Does not permit duplicates.
+	 * If this is a subtyped serializer, the first serializer registered will also be registered with key {@link PlaceboJsonReloadListener#DEFAULT}.
+	 * @param id The ID of the serializer.  If you do not support subtypes, use {@link PlaceboJsonReloadListener#DEFAULT}.
+	 * @param serializer The serializer being registered.
+	 */
 	public final void registerSerializer(ResourceLocation id, SerializerBuilder<V> serializer) {
 		if (this.subtypes) {
 			if (this.serializers.containsKey(id)) throw new RuntimeException("Attempted to register a " + this.path + " serializer with id " + id + " but one already exists!");
@@ -186,20 +209,64 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 	public static boolean checkAndLogEmpty(JsonElement e, ResourceLocation id, String type, Logger logger) {
 		String s = e.toString();
 		if (s.isEmpty() || s.equals("{}")) {
-			logger.debug("Ignoring {} item with id {} as it is empty.", type, id);
+			logger.error("Ignoring {} item with id {} as it is empty.  Please switch to a condition-false json instead of an empty one.", type, id);
 			return false;
 		}
 		return true;
 	}
 
-	public static boolean checkConditions(JsonElement e, ResourceLocation id, String type, Logger logger) {
-		if (e.isJsonObject() && !CraftingHelper.processConditions(e.getAsJsonObject(), "conditions")) {
+	/**
+	 * Checks the conditions on a Json, and returns true if they are met.
+	 * @param e The Json being checked.
+	 * @param id The ID of that json.
+	 * @param type The type of the json, for logging.
+	 * @param logger The logger to log to.
+	 * @param context The context object used for resolving conditions.
+	 * @return True if the item's conditions are met, false otherwise.
+	 */
+	public static boolean checkConditions(JsonElement e, ResourceLocation id, String type, Logger logger, IContext context) {
+		if (e.isJsonObject() && !CraftingHelper.processConditions(e.getAsJsonObject(), "conditions", context)) {
 			logger.debug("Skipping loading {} item with id {} as it's conditions were not met", type, id);
 			return false;
 		}
 		return true;
 	}
 
+	/**
+	 * @return The context object held in this listener, or {@link IContext.EMPTY} if it is unavailable.
+	 */
+	protected final ICondition.IContext getContext() {
+		return this.context.get() != null ? this.context.get() : IContext.EMPTY;
+	}
+
+	/**
+	 * Registers this listener to the event bus as is appropriate.
+	 * This should be called for ALL listeners from common setup.
+	 */
+	public void registerToBus() {
+		if (this.synced) registerForSync(this);
+		MinecraftForge.EVENT_BUS.addListener(this::addReloader);
+	}
+
+	private final void registerForSync(PlaceboJsonReloadListener<?> listener) {
+		if (!listener.synced) throw new RuntimeException("Attempted to register the non-synced JSON Reload Listener " + listener.path + " as a synced listener!");
+		synchronized (SYNC_REGISTRY) {
+			if (SYNC_REGISTRY.containsKey(listener.path)) throw new RuntimeException("Attempted to register the JSON Reload Listener for syncing " + listener.path + " but one already exists!");
+			SYNC_REGISTRY.put(listener.path, listener);
+			MinecraftForge.EVENT_BUS.addListener(listener::sync);
+		}
+	}
+
+	private final void addReloader(AddReloadListenerEvent e) {
+		e.addListener(this);
+		this.context = new WeakReference<>(e.getConditionContext());
+	}
+
+	/**
+	 * An interface supporting ID's and Serializers with generic types.
+	 *
+	 * @param <V> This
+	 */
 	public static interface TypeKeyed<V extends TypeKeyed<V>> {
 		void setId(ResourceLocation id);
 
@@ -210,6 +277,11 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		SerializerBuilder<V>.Serializer getSerializer();
 	}
 
+	/**
+	 * Intrusive base implementation of {@link TypeKeyed}
+	 *
+	 * @param <V> This
+	 */
 	public static abstract class TypeKeyedBase<V extends TypeKeyed<V>> implements TypeKeyed<V> {
 		protected ResourceLocation id;
 		protected SerializerBuilder<V>.Serializer serializer;
@@ -237,6 +309,10 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		}
 	}
 
+	/**
+	 * Begins the sync for a specific listener.
+	 * @param path The path of the listener being synced.
+	 */
 	public static void initSync(String path) {
 		SYNC_REGISTRY.computeIfPresent(path, (k, v) -> {
 			v.staged.clear();
@@ -244,6 +320,13 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		});
 	}
 
+	/**
+	 * Write an item (with the same type as the listener) to the network.
+	 * @param <V> The type of item being written.
+	 * @param path The path of the listener.
+	 * @param value The value being written.
+	 * @param buf The buffer being written to.
+	 */
 	public static <V extends TypeKeyed<V>> void writeItem(String path, V value, FriendlyByteBuf buf) {
 		SYNC_REGISTRY.computeIfPresent(path, (k, v) -> {
 			ResourceLocation serId = v.serializers.inverse().get(value.getSerializer());
@@ -253,6 +336,14 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		});
 	}
 
+	/**
+	 * Reads an item from the network, via the listener's serializers.
+	 * @param <V> The type of item being read.
+	 * @param path The path of the listener.
+	 * @param key The key of the item being read (not the serializer key).
+	 * @param buf The buffer being read from.
+	 * @return An object of type V as deserialized from the network.
+	 */
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	public static <V extends TypeKeyed<V>> V readItem(String path, ResourceLocation key, FriendlyByteBuf buf) {
 		var listener = SYNC_REGISTRY.get(path);
@@ -264,16 +355,27 @@ public abstract class PlaceboJsonReloadListener<V extends TypeKeyed<V>> extends 
 		return v;
 	}
 
+	/**
+	 * Stages an item to a listener.
+	 * @param <V> The type of the item being staged.
+	 * @param path The path of the listener.
+	 * @param value The object being staged.
+	 */
 	@SuppressWarnings({ "unchecked", "rawtypes" })
-	public static <V extends TypeKeyed<V>> void acceptItem(String path, ResourceLocation key, V value) {
+	public static <V extends TypeKeyed<V>> void acceptItem(String path, V value) {
 		SYNC_REGISTRY.computeIfPresent(path, (k, v) -> {
-			((Map) v.staged).put(key, value);
+			((Map) v.staged).put(value.getId(), value);
 			return v;
 		});
 	}
 
+	/**
+	 * Ends the sync for a specific listener.
+	 * This will delete current data, push staged data to live, and call the appropriate methods for reloading.
+	 * @param path The path of the listener.
+	 */
 	@SuppressWarnings({ "unchecked", "rawtypes" })
-	public static <V extends TypeKeyed<V>> void endSync(String path) {
+	public static void endSync(String path) {
 		SYNC_REGISTRY.computeIfPresent(path, (k, v) -> {
 			v.beginReload();
 			v.staged.forEach(((PlaceboJsonReloadListener) v)::register);